Fix refreshing: not either/or.
Firstly, chests with a reset value of 1 or 0 are considered ready to be
refreshed.

Secondly, instead of only refreshing the chest OR notify that there's an
upcoming refresh (and marking it for a refresh), it now does both. This
means a chest can refresh and start the cooldown for another refresh
immediately.