This fix doesn't actually fix it.
I'm not sure why; it doesn't seem to send the actual packet, implying
that we might need to defer it until later.

That said, the NeoForge hook likewise doesn't re-send the actual block
entity tag as expected -- or rather, the `onDataPacket` function is
never called.

Is this because the block entity on the client-side doesn't exist?