Improve null handling even more.
This should finally resolve issues where server tick events calling
`doDecay`, etc, cause null pointer issues where the server is null or
the return of `MinecraftServer::overworld` is null.

I'm still convinced that these "issues" are simply masking the fact that
the server is in an errored state, or that these server tick events are
being called before the server has been set up -- all of which make no
sense, as the overworld is the first level created, and as such is
created in the instantiation of the server itself.