Clear openers when refreshing a chest.
This is for #205, however it will not resolve the main issue that people
have, which is that Lootr containers aren't constantly polled, but are
only refreshed when they are opened.

This means that a chest might be ready to refresh but has not yet done
so, displaying the old texture.