Potentially better handling for #313.
Error messages when attempting to save including the block position and
dimension instead of just straight up crashing.

Properly check for a null `tileId` when loading and generate a new
`tileId` if it's null, instead of if it's just not loaded properly. I'm
not sure if it's ever possible for this to take place, but it's worth
it.

Finally, if a player opens a container which doesn't have a generated
`tileId` for whatever reason, inform them of that fact.