Store block type info.
This is specifically for decoding data files at a later date, presumably
when Lootr is no longer available, in order to potentially convert them
back to their "originals".

Obviously it won't support modded containers as extensively, but it
should allow for a generalization, and modded containers should natively
support the removal of Lootr.

Exceptions: it'll require specific code in order to transform
inventories back into chests.