Null handling for entities/block entities.
It shouldn't actually be possible for these to be null, but in rare
circumstances it seems like the PacketCloseCart can be sent with an
entity id that resolves to null on the client-side while still being
present on the server side.

Adding breakpoints to this becomes difficult, implying that it's a race
condition with the `startSeenByPlayer` firing a packet before the entity
has been properly created on the client-side.

It shouldn't be an issue but I've made a note just in case.